refactor: 사용자용 예약 서비스와 관리자용 예약 서비스 구현체를 통합하고, 인터페이스로 분리
